requirements.txt
django
python3
djangorestframework-simplejwt
channels
pandas
numpy